+++ DelvUI/changelog.md
@@ -9,6 +9,7 @@
     
 - Moved "Experience Bar", "GCD Indicator", "Pull Timer", "Limit Break" and "MP Ticker" from "Misc" to the new "Other Elements" section.
 - Moved "Fonts" from "Misc" to the new "Customization" section.
+- Fixed targeted unit Nameplate not drawing on top of other Nameplates.
 
 # 1.7.1.1
 - Fixed Scholar's Fairy Gauge Bar not working.